refactor mockRouter
